Rag_Chat/backend/requirements.txt
# Django & DRF
Django>=4.2.0,<5.0.0
djangorestframework>=3.14.0
django-cors-headers>=4.2.0

# Authentication
djangorestframework-simplejwt>=5.3.0

# Frontend
streamlit>=1.29.0

# Environment variables
python-dotenv>=1.0.0
django-environ>=0.10.0

# Production
gunicorn>=21.2.0
whitenoise>=6.5.0

# Utils
requests>=2.31.0
redis>=5.0.1
celery>=5.3.6
django-celery-beat>=2.5.0
python-jose>=3.3.0

# Testing & CI/CD
pytest>=7.4.0
pytest-django>=4.5.2
flake8>=6.1.0
coverage>=7.3.0
black>=23.12.1
isort>=5.13.2

# LangChain
langchain>=0.3.18,<0.4.0
langchain-community>=0.3.17,<0.4.0
langchain-core>=0.3.34,<0.4.0
langchain-openai>=0.3.4
langchain-google-genai>=1.0.7
chromadb==0.4.22
pandas>=2.1.4
openpyxl>=3.1.2
